feat(inventory): inventory service

+++ src/main/java/com/futurenbetter/saas/modules/inventory/entity/InventoryInvoiceDetail.java
@@ -41,7 +41,7 @@ public class InventoryInvoiceDetail {
     Integer inputQuantity;
 
     @Column(name = "converted_quantity", nullable = false)
-    Integer convertedQuantity;
+    Double convertedQuantity;
 
     @Column(name = "unit_price", nullable = false)
     Double unitPrice;
